-Refactors the sensor modifications into their own classes.
