+++ app/src/main/java/com/mrchandler/disableprox/util/ProUtil.java
@@ -48,6 +48,6 @@ public final class ProUtil {
     }
 
     private static SharedPreferences getSharedPreferences(Context context) {
-        return context.getSharedPreferences(Constants.PREFS_FILE_NAME, Context.MODE_WORLD_READABLE);
+        return context.getSharedPreferences(Constants.PREFS_FILE_NAME, Context.MODE_PRIVATE);
     }
 }
